Massive update
Added the generic clustering mechanism.
Now you can type your own type of cluster_item, the system will recognize your type and permitsm, the clustering of a list of your type and the serialization/deserialization through a SQL db of the items likes a single dataset.

A simple but not complete front end is added to show the results of the clustering process.